requirements-serve.txt
torch==2.11.0
torchvision==0.26.0
pillow
mlserver
mlserver-mlflow
